+++ book.rb
@@ -1,6 +1,8 @@
+require_relative './person'
 class Book
   attr_reader :borrowings
   attr_accessor :title, :author
+
   def initialize(title, author)
     @title = title
     @author = author
