+++ classroom.rb
@@ -1,11 +1,13 @@
+require_relative './person'
 class Classroom
   attr_accessor :label
   attr_reader :students
+
   def initialize(label)
     @label = label
     @students = []
   end
-  
+
   def add_student(student)
     @students.push(student)
     student.class1 = self
