+++ nameable.rb
@@ -1,3 +1,4 @@
+require_relative './person'
 class Nameable
   def correct_name
     raise NotImplementedError, "#{self.class} has not implemented method '#{__method__}'"
