+++ rental.rb
@@ -1,0 +1,15 @@
+require_relative './person'
+class Rental
+  attr_reader :person, :book
+  attr_accessor :date
+
+  def initialize(date, person, book)
+    @date = date
+
+    @person = person
+    person.borrowings << self
+
+    @book = book
+    person.borrowings << self
+  end
+end
